6.3.1 - Small additions and code cleanup Now TARGETTING .NET 5.0

+++ InterlockLedger.Tags/Core/Metadata/DataModel.cs
@@ -265,7 +265,7 @@ namespace InterlockLedger.Tags
                     DataFields = s.DecodeTagArray<ILTagDataField>()?.Select(t => t.Value),
                     Indexes = s.DecodeTagArray<ILTagDataIndex>()?.Select(t => t.Value),
                     PayloadName = s.DecodeString(),
-                    Version = s.HasBytes() ? s.DecodeUShort() : (ushort)1,
+                    Version = s.HasBytes() ? s.DecodeUShort() : 1,
                     Description = s.HasBytes() ? s.DecodeString().TrimToNull() : null
                 };
             });
